refactor(icu_parser): consolidate intl polyfill
