+++ packages/icu-messageformat-parser/src/ast.rs
@@ -1,8 +1,10 @@
-use super::js_intl::*;
 use serde::ser::{SerializeMap, SerializeStruct};
 use serde::{Serialize, Serializer};
 use serde_repr::Serialize_repr;
 use std::fmt;
+
+use crate::intl::date_time_format_options::JsIntlDateTimeFormatOptions;
+use crate::intl::number_format_options::JsIntlNumberFormatOptions;
 
 /// The type of an error that occurred while building an AST.
 #[derive(Clone, Debug, Eq, PartialEq, Serialize_repr)]
@@ -393,7 +395,7 @@ pub struct PluralOrSelectOption<'s> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::js_intl::JsIntlNumberFormatOptions;
+    use crate::intl::number_format_options::JsIntlNumberFormatOptions;
     use serde_json::json;
 
     #[test]
