+++ packages/icu-messageformat-parser/src/intl/locale.rs
@@ -1,0 +1,3 @@
+pub struct Locale {
+
+}
