+++ packages/icu-messageformat-parser/src/lib.rs
@@ -1,7 +1,7 @@
 mod ast;
-mod js_intl;
 mod parser;
 mod pattern_syntax;
+mod intl;
 
 pub use ast::{Ast, AstElement, Error, Position, Span};
 pub use parser::{Parser, ParserOptions};
